requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
pyodbc
azure-identity
pandas
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-storage
azure-mgmt-resource
azure.mgmt.monitor
